Primeira grande refatoração
App renomeado para lutabrawar
Filtros movidos para um arquivo próprio
Adição de um novo filtro gruped_by
