+++ lutabrawar/wsgi.py
@@ -1,4 +1,4 @@
-from app import app as application
+from lutabrawar import app as application
 
 if __name__ == '__main__':
     application.run()
